corrección de la caga que quedo anoche
lo mismo

+++ ms-componentes/src/main/java/com/pcbuilder/ms_componentes/service/ComponenteService.java
@@ -40,7 +40,9 @@ public class ComponenteService {
     }
 
     public void eliminar(Long id) {
-        buscarEntidadPorId(id);
+        if (!repo.existsById(id)) {
+            throw new RecursoNoEncontradoException("El componente con ID " + id + " no existe.");
+        }
         repo.deleteById(id);
     }
